refactor context into reusable hook
